[Prakash] - Added the ability for the service layer to store the employee payroll data in a list

+++ src/main/java/com/bridgelabz/employeepayrollh2db/models/EmployeePayrollDTO.java
@@ -18,7 +18,7 @@ public class EmployeePayrollDTO {
 		this.name = name;
 	}
 
-	public double getSalary() {
+	public long getSalary() {
 		return salary;
 	}
