+++ src/main/java/com/bridgelabz/employeepayrollh2db/models/EmployeePayrollDetails.java
@@ -29,7 +29,7 @@ public class EmployeePayrollDetails {
 		this.name = name;
 	}
 
-	public double getSalary() {
+	public long getSalary() {
 		return salary;
 	}
